io_file implementation

+++ libs/array.h
@@ -120,7 +120,7 @@ ARRAY_DEF int arr_contains(const Arr *arr, void *src) {
 
 ARRAY_DEF int arr_index(const Arr *arr, void *src) {
     ARRAY_CHECK_NOTNULL(arr);
-    return ((char *) src - (char *) arr->data) / arr->msize;
+    return (int) (((char *) src - (char *) arr->data) / arr->msize);
 }
 
 ARRAY_DEF void *arr_pop(Arr *arr) {
